+++ variation-commons-core/src/main/java/uk/ac/ebi/eva/commons/core/models/IFeatureCoordinates.java
@@ -25,8 +25,8 @@ public interface IFeatureCoordinates {
 
     String getChromosome();
 
-    int getStart();
+    long getStart();
 
-    int getEnd();
+    long getEnd();
 
 }
